helper function to print better string moved to priting file
